Thanks to @andinux: fixed an inifinite loop with thickness divider on subviews adjustments and infinite loop bug while resizing with priority value < 0.5
#Thickness Bug
Delta adjustment needed because the original thickness of a NSSplitView
created with the xib is 1 and the subviews are sized consequently so
the thickness is changed programmatically, before resizing the subviews
proportionally I have to consider this delta adjustment. now this
adjustment is calculated in the dividerThickness setter, but it works
just when the original thickness used to create the initial subviews
frame is equal to 1 (like in the xib) another way to calculate this
delta (stronger but more expensive) may be to caulculate every time in
"applyProportionalResizeFromOldSize:" the total size of the subviews
obtaining the delta adjustment with totalSize-oldsize.

#Infinite Loop Bug:
avoid loop (for example if the remaining value is 1 and no one subview
has a proportion greather that 0.5, each subviewDelta will be 0 and we
have a loop)